fix: normalize algorithm name for TOTP Auth URI generation
Ensure the algorithm name is compatible with the otpauth URI specification by removing dashes from the algorithm name.

BREAKING CHANGE: Technically, we are changing the algorithm param that is returned from getTOTPAuthUri and that could be a breaking change for some people, but it shouldn't be a breaking change for most (any). So go ahead and update this without concern.